Fix crash in libecho.c on win64 build
Return value of _findfirst is intptr_t, see https://msdn.microsoft.com/en-us/library/zyzxfzac.aspx

+++ src/util/windows/libecho.c
@@ -38,7 +38,7 @@ main(int argc, char *argv[])
 void
 echo_files(char *prefix, char *f)
 {
-    long ff;
+    intptr_t ff;
     struct _finddata_t fdt;
     char *slash;
     char filepath[256];
